+++ include/build.h
@@ -4,5 +4,5 @@
 
 #pragma once
 
-#define __BUILD_NUMBER "7001"
+#define __BUILD_NUMBER "7054"